+++ src/aql/llm_aql_handler.cpp
@@ -1716,7 +1716,7 @@ void LLMAQLHandler::logAnnotations(const std::vector<AQLAnnotation> &annotations
 
     constexpr std::size_t MAX_PREVIEW = 100;
     std::string preview
-        = query_preview.size() > MAX_PREVIEW ? query_preview.substr(0, MAX_PREVIEW) + "..." : query_preview;
+        = static_cast<int>(query_preview.size()) > MAX_PREVIEW ? query_preview.substr(0, MAX_PREVIEW) + "..." : query_preview;
 
     std::ostringstream warn_msg = {};
     warn_msg << function_name << " produced " <<static_cast<int>(annotations.size()) << " potential syntax issue(s) for query \""
@@ -1733,7 +1733,7 @@ std::string LLMAQLHandler::translateNLToAQL(const std::string &nl_query, const s
     sanitizePromptInput(schema_context, "schema_context", impl_->validation_limits_.max_schema_context_length);
 
     spdlog::debug("NL-to-AQL: Starting translation for query: {}", 
-                  nl_query.size() > 100 ? nl_query.substr(0, 100) + "..." : nl_query);
+                  static_cast<int>(nl_query.size()) > 100 ? nl_query.substr(0, 100) + "..." : nl_query);
 
     const TranslationValidationMode mode = impl_->validation_mode_;
     const size_t max_attempts = getConfiguredRetryAttempts(mode, impl_->config_.validation_config);
@@ -2202,7 +2202,7 @@ std::string LLMAQLHandler::translateNLToAQLWithExamples(const std::string &nl_qu
             }
 
             spdlog::debug("translateNLToAQLWithExamples: injected {} examples for query \"{}\"", injected_count,
-                          nl_query.size() > 60 ? nl_query.substr(0, 60) + "..." : nl_query);
+                          static_cast<int>(nl_query.size()) > 60 ? nl_query.substr(0, 60) + "..." : nl_query);
 
             return aql_query;
