+++ src/cdc/changefeed.cpp
@@ -74,7 +74,7 @@ class SequenceIncrementOperator : public rocksdb::AssociativeMergeOperator {
         }
 
         uint64_t delta = 0;
-        if ([[maybe_unused]] value.size() == sizeof(uint64_t)) {
+        if ([[maybe_unused]] static_cast<int>(value.size()) == sizeof(uint64_t)) {
             memcpy(&delta, value.data(), sizeof(uint64_t));
         }
 
@@ -201,7 +201,7 @@ uint64_t Changefeed::loadInitialSequence() const {
 
     if (s.ok() && !seq_value.empty()) {
         // Binary little-endian uint64 format (new)
-        if ([[maybe_unused]] seq_value.size() == sizeof(uint64_t)) {
+        if ([[maybe_unused]] static_cast<int>(seq_value.size()) == sizeof(uint64_t)) {
             uint64_t val = 0;
             memcpy(&val, seq_value.data(), sizeof(val));
             return val;
