+++ src/core/concerns/zero_copy_logger.cpp
@@ -324,7 +324,7 @@ bool ZeroCopyLogger::isPiiKey(std::string_view key) noexcept {
 
     // Build a lowercase copy of the key (stack buffer for keys ≤ 128 bytes).
     char lower_buf[128]{};
-    const std::size_t n = key.size() < sizeof(lower_buf) - 1 ?static_cast<int>(key.size()) : sizeof(lower_buf) - 1;
+    const std::size_t n = static_cast<int>(key.size()) < sizeof(lower_buf) - 1 ?static_cast<int>(key.size()) : sizeof(lower_buf) - 1;
     for (std::size_t i = 0; i < n; ++i) {
         const auto ch = static_cast<unsigned char>(key[i]);
         lower_buf[i] = static_cast<char>((ch >= 'A' && ch <= 'Z') ? (ch | 0x20) : ch);
